attached_assets/Pasted-Server-Error-Error-Element-type-is-invalid-expected-a-string-for-built-in-components-or-a-class-1748443993688.txt

Server Error
Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

This error happened while generating the page. Any console logs will be displayed in the terminal window.
Call Stack
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6043:9)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderChildrenArray
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6211:7)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6141:7)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderHostElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5642:3)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5952:5)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderForwardRef
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5859:5)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6005:11)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderForwardRef
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5859:5)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6005:11)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderChildrenArray
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6211:7)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6141:7)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderHostElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5642:3)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5952:5)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderChildrenArray
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6211:7)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6141:7)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderContextProvider
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5920:3)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6017:11)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderForwardRef
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5859:5)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6005:11)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderChildrenArray
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6211:7)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6141:7)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
renderNode
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6259:12)
renderHostElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5642:3)
renderElement
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (5952:5)
renderNodeDestructiveImpl
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6104:11)
renderNodeDestructive
file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.browser.development.js (6076:14)
